Move participant count types to analysis
