+++ src/analysis/types.ts
